+++ packages/react/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @lezzform/react
 
+## 0.2.20
+
+### Patch Changes
+
+- improve dateRangePicker
+- Updated dependencies
+  - @lezzform/types@0.0.20
+
+## 0.2.19
+
+### Patch Changes
+
+- remove date range picker default value
+
+## 0.2.18
+
+### Patch Changes
+
+- init date range picker
+- Updated dependencies
+  - @lezzform/types@0.0.19
+
 ## 0.2.17
 
 ### Patch Changes